User functionality fully cleaned and working
