+++ .idea/jsLibraryMappings.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="JavaScriptLibraryMappings">
-    <file url="file://$PROJECT_DIR$" libraries="{BTSwebsite/node_modules}" />
-    <file url="file://$PROJECT_DIR$/routes/index.js" libraries="{WebGL}" />
+    <file url="file://$PROJECT_DIR$/routes/users.js" libraries="{WebGL}" />
   </component>
 </project>
