+++ views/register.hbs
@@ -36,7 +36,7 @@
         <div class="page-header">
             <h1>My Profile</h1>
         </div>
-        <form method="post" role="form" action="/register" class="login-form form-horizontal">
+        <form method="post" role="form" action="/users/register" class="login-form form-horizontal">
             <input name="_csrf" type="hidden"/>
             <div class="form-group">
                 <label class="col-sm-4">User Name</label>
